+++ sdk/required_services.d.ts
@@ -15,12 +15,14 @@
  */
 
 import {google as adsAccount} from 'service/ads/account-service';
+import {google as adsDateRange} from 'service/ads/ads-date-range-service';
 import {google as adsConversionTracking} from 'service/ads/conversion-tracking-service';
 import {google as accountStatus} from 'service/partner/account-status-service';
 import {google as authentication} from 'service/partner/authentication-service';
 import {google as business} from 'service/partner/business-service';
 import {google as campaign} from 'service/partner/campaign-service';
 import {google as partnerConversionTracking} from 'service/partner/conversion-tracking-service';
+import {google as partnerDateRange} from 'service/partner/partner-date-range-service';
 import {google as partnerTermsAndConditions} from 'service/partner/terms-and-conditions-service';
 /**
  * Namespace for all the services to communicate between the Ads app and the
@@ -46,6 +48,9 @@ export declare namespace google.ads.integration.service {
         accountStatus.ads.integration.service.partner.AccountStatusService;
 
     campaignService?: campaign.ads.integration.service.partner.CampaignService;
+
+    partnerDateRangeService?: partnerDateRange.ads.integration.service.partner
+        .PartnerDateRangeService;
   }
 
   /**
@@ -56,5 +61,7 @@ export declare namespace google.ads.integration.service {
     conversionTrackingIdService?: adsConversionTracking.ads.integration.service
         .ads.ConversionTrackingIdService;
     accountService?: adsAccount.ads.integration.service.ads.AccountService;
+    adsDateRangeService?:
+        adsDateRange.ads.integration.service.ads.AdsDateRangeService;
   }
 }